+++ src/vec/vec/impls/seq/seqcuda/veccuda2.cu
@@ -53,7 +53,7 @@ PetscErrorCode VecCUDAAllocateCheck(Vec v)
     if (option_set) {
       v->minimum_bytes_pinned_memory = pinned_memory_min;
 #if defined(PETSC_HAVE_ISINF)
-      if isinf(pinned_memory_min) v->minimum_bytes_pinned_memory = -1;
+      if (isinf(pinned_memory_min)) v->minimum_bytes_pinned_memory = (size_t)-1;
 #endif
     }
     ierr = PetscOptionsEnd();CHKERRQ(ierr);
